Remove `DescriptorDevice` trait
Also move the implementation to `descriptor.rs`.
